--- shell/app/electron_main.h
@@ -1,10 +1,0 @@
-// Copyright (c) 2013 GitHub, Inc.
-// Use of this source code is governed by the MIT license that can be
-// found in the LICENSE file.
-
-#ifndef ELECTRON_SHELL_APP_ELECTRON_MAIN_H_
-#define ELECTRON_SHELL_APP_ELECTRON_MAIN_H_
-
-#include "content/public/app/content_main.h"
-
-#endif  // ELECTRON_SHELL_APP_ELECTRON_MAIN_H_
