+++ shell/app/electron_main_win.cc
@@ -1,8 +1,13 @@
-// Copyright (c) 2013 GitHub, Inc.
+// Copyright (c) 2022 Slack Technologies, Inc.
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "shell/app/electron_main.h"
+#include <windows.h>  // windows.h must be included first
+
+#include <atlbase.h>  // ensures that ATL statics like `_AtlWinModule` are initialized (it's an issue in static debug build)
+#include <shellapi.h>
+#include <shellscalingapi.h>
+#include <tchar.h>
 
 #include <algorithm>
 #include <cstdlib>
@@ -11,100 +16,41 @@
 #include <utility>
 #include <vector>
 
-#if defined(OS_POSIX)
-#include <sys/stat.h>
-#include "base/ignore_result.h"
-#endif
-
-#if defined(OS_WIN)
-#include <windows.h>  // windows.h must be included first
-
-#include <atlbase.h>  // ensures that ATL statics like `_AtlWinModule` are initialized (it's an issue in static debug build)
-#include <shellapi.h>
-#include <shellscalingapi.h>
-#include <tchar.h>
-
+#include "base/at_exit.h"
 #include "base/environment.h"
+#include "base/i18n/icu_util.h"
 #include "base/process/launch.h"
 #include "base/strings/utf_string_conversions.h"
 #include "base/win/windows_version.h"
 #include "components/browser_watcher/exit_code_watcher_win.h"
 #include "components/crash/core/app/crash_switches.h"
 #include "components/crash/core/app/run_as_crashpad_handler_win.h"
+#include "content/public/app/content_main.h"
 #include "content/public/app/sandbox_helper_win.h"
+#include "electron/buildflags/buildflags.h"
+#include "electron/fuses.h"
 #include "sandbox/win/src/sandbox_types.h"
 #include "shell/app/command_line_args.h"
 #include "shell/app/electron_main_delegate.h"
-#include "third_party/crashpad/crashpad/util/win/initial_client_data.h"
-
-#elif defined(OS_LINUX)  // defined(OS_WIN)
-#include <unistd.h>
-#include <cstdio>
-#include "base/base_switches.h"
-#include "base/command_line.h"
-#include "content/public/app/content_main.h"
-#include "shell/app/electron_main_delegate.h"  // NOLINT
-#else                                          // defined(OS_LINUX)
-#include <mach-o/dyld.h>
-#include <unistd.h>
-#include <cstdio>
-#include "shell/app/electron_library_main.h"
-#endif  // defined(OS_MAC)
-
-#include "base/at_exit.h"
-#include "base/i18n/icu_util.h"
-#include "electron/buildflags/buildflags.h"
-#include "electron/fuses.h"
 #include "shell/app/node_main.h"
 #include "shell/common/electron_command_line.h"
 #include "shell/common/electron_constants.h"
-
-#if defined(HELPER_EXECUTABLE) && !defined(MAS_BUILD)
-#include "sandbox/mac/seatbelt_exec.h"  // nogncheck
-#endif
+#include "third_party/crashpad/crashpad/util/win/initial_client_data.h"
 
 namespace {
 
-#if defined(OS_WIN)
 // Redefined here so we don't have to introduce a dependency on //content
 // from //electron:electron_app
 const char kUserDataDir[] = "user-data-dir";
 const char kProcessType[] = "type";
-#endif
 
 ALLOW_UNUSED_TYPE bool IsEnvSet(const char* name) {
-#if defined(OS_WIN)
   size_t required_size;
   getenv_s(&required_size, nullptr, 0, name);
   return required_size != 0;
-#else
-  char* indicator = getenv(name);
-  return indicator && indicator[0] != '\0';
-#endif
 }
-
-#if defined(OS_POSIX)
-void FixStdioStreams() {
-  // libuv may mark stdin/stdout/stderr as close-on-exec, which interferes
-  // with chromium's subprocess spawning. As a workaround, we detect if these
-  // streams are closed on startup, and reopen them as /dev/null if necessary.
-  // Otherwise, an unrelated file descriptor will be assigned as stdout/stderr
-  // which may cause various errors when attempting to write to them.
-  //
-  // For details see https://github.com/libuv/libuv/issues/2062
-  struct stat st;
-  if (fstat(STDIN_FILENO, &st) < 0 && errno == EBADF)
-    ignore_result(freopen("/dev/null", "r", stdin));
-  if (fstat(STDOUT_FILENO, &st) < 0 && errno == EBADF)
-    ignore_result(freopen("/dev/null", "w", stdout));
-  if (fstat(STDERR_FILENO, &st) < 0 && errno == EBADF)
-    ignore_result(freopen("/dev/null", "w", stderr));
-}
-#endif
 
 }  // namespace
-
-#if defined(OS_WIN)
 
 namespace crash_reporter {
 extern const char kCrashpadProcess[];
@@ -291,74 +237,3 @@ int APIENTRY wWinMain(HINSTANCE instance, HINSTANCE, wchar_t* cmd, int) {
   electron::ElectronCommandLine::Init(arguments.argc, arguments.argv);
   return content::ContentMain(std::move(params));
 }
-
-#elif defined(OS_LINUX)  // defined(OS_WIN)
-
-int main(int argc, char* argv[]) {
-  FixStdioStreams();
-
-#if BUILDFLAG(ENABLE_RUN_AS_NODE)
-  if (electron::fuses::IsRunAsNodeEnabled() && IsEnvSet(electron::kRunAsNode)) {
-    base::i18n::InitializeICU();
-    base::AtExitManager atexit_manager;
-    return electron::NodeMain(argc, argv);
-  }
-#endif
-
-  electron::ElectronMainDelegate delegate;
-  content::ContentMainParams params(&delegate);
-  electron::ElectronCommandLine::Init(argc, argv);
-  params.argc = argc;
-  params.argv = const_cast<const char**>(argv);
-  base::CommandLine::Init(params.argc, params.argv);
-  // TODO(https://crbug.com/1176772): Remove when Chrome Linux is fully migrated
-  // to Crashpad.
-  base::CommandLine::ForCurrentProcess()->AppendSwitch(
-      ::switches::kEnableCrashpad);
-  return content::ContentMain(std::move(params));
-}
-
-#else  // defined(OS_LINUX)
-
-int main(int argc, char* argv[]) {
-  FixStdioStreams();
-
-#if BUILDFLAG(ENABLE_RUN_AS_NODE)
-  if (electron::fuses::IsRunAsNodeEnabled() && IsEnvSet(electron::kRunAsNode)) {
-    return ElectronInitializeICUandStartNode(argc, argv);
-  }
-#endif
-
-#if defined(HELPER_EXECUTABLE) && !defined(MAS_BUILD)
-  uint32_t exec_path_size = 0;
-  int rv = _NSGetExecutablePath(NULL, &exec_path_size);
-  if (rv != -1) {
-    fprintf(stderr, "_NSGetExecutablePath: get length failed\n");
-    abort();
-  }
-
-  auto exec_path = std::make_unique<char[]>(exec_path_size);
-  rv = _NSGetExecutablePath(exec_path.get(), &exec_path_size);
-  if (rv != 0) {
-    fprintf(stderr, "_NSGetExecutablePath: get path failed\n");
-    abort();
-  }
-  sandbox::SeatbeltExecServer::CreateFromArgumentsResult seatbelt =
-      sandbox::SeatbeltExecServer::CreateFromArguments(exec_path.get(), argc,
-                                                       argv);
-  if (seatbelt.sandbox_required) {
-    if (!seatbelt.server) {
-      fprintf(stderr, "Failed to create seatbelt sandbox server.\n");
-      abort();
-    }
-    if (!seatbelt.server->InitializeSandbox()) {
-      fprintf(stderr, "Failed to initialize sandbox.\n");
-      abort();
-    }
-  }
-#endif  // defined(HELPER_EXECUTABLE) && !defined(MAS_BUILD)
-
-  return ElectronMain(argc, argv);
-}
-
-#endif  // defined(OS_MAC)
